Tableau Dashboard
The Dashboard includes information about the current performance of the department (i.e. number of new cases filed and patents issued YTD vs the same of the previous year), the current workload of the department (i.e. number of current new applications to be filed and pending cases.) It also includes the performance of the working attorney (i.e. their current workload, avg time spent on a patent case, and avg number of rejections received on their cases.) The management can utilize these statistics to allocate the work when new cases come in.